@@ -1,19 +1,19 @@
 name: esc_pos_printer
 description: The library allows to print receipts using an ESC/POS thermal WiFi printer.
-version: 4.0.3
+version: 4.0.3+1
 homepage: https://github.com/andrey-ushakov/esc_pos_printer
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  esc_pos_utils: ^1.0.0
-  # esc_pos_utils:
-  #   path: ../esc_pos_utils
-  charset_converter: ^1.0.3
-  image: ^2.1.4
+  charset_converter: ^2.0.0
+  image: ^3.0.2
+  esc_pos_utils:
+    git:
+      url: https://github.com/MasterShu/esc_pos_utils
 
 dev_dependencies:
   flutter_test:
