@@ -1,7 +1,7 @@
 /*
  * esc_pos_printer
  * Created by Andrey Ushakov
- * 
+ *
  * Copyright (c) 2019-2020. All rights reserved.
  * See LICENSE for distribution and usage details.
  */
@@ -21,13 +21,13 @@ class NetworkPrinter {
 
   final PaperSize _paperSize;
   final CapabilityProfile _profile;
-  String _host;
-  int _port;
-  Generator _generator;
-  Socket _socket;
+  String? _host;
+  int? _port;
+  late Generator _generator;
+  late Socket _socket;
 
-  int get port => _port;
-  String get host => _host;
+  int? get port => _port;
+  String? get host => _host;
   PaperSize get paperSize => _paperSize;
   CapabilityProfile get profile => _profile;
 
@@ -45,7 +45,7 @@ class NetworkPrinter {
   }
 
   /// [delayMs]: milliseconds to wait after destroying the socket
-  void disconnect({int delayMs}) async {
+  void disconnect({int? delayMs}) async {
     _socket.destroy();
     if (delayMs != null) {
       await Future.delayed(Duration(milliseconds: delayMs), () => null);
@@ -58,12 +58,12 @@ class NetworkPrinter {
   }
 
   void text(
-    String text, {
-    PosStyles styles = const PosStyles(),
-    int linesAfter = 0,
-    bool containsChinese = false,
-    int maxCharsPerLine,
-  }) {
+      String text, {
+        PosStyles styles = const PosStyles(),
+        int linesAfter = 0,
+        bool containsChinese = false,
+        int? maxCharsPerLine,
+      }) {
     _socket.add(_generator.text(text,
         styles: styles,
         linesAfter: linesAfter,
@@ -75,7 +75,7 @@ class NetworkPrinter {
     _socket.add(_generator.setGlobalCodeTable(codeTable));
   }
 
-  void setGlobalFont(PosFontType font, {int maxCharsPerLine}) {
+  void setGlobalFont(PosFontType font, {int? maxCharsPerLine}) {
     _socket
         .add(_generator.setGlobalFont(font, maxCharsPerLine: maxCharsPerLine));
   }
@@ -100,7 +100,7 @@ class NetworkPrinter {
     _socket.add(_generator.cut(mode: mode));
   }
 
-  void printCodeTable({String codeTable}) {
+  void printCodeTable({String? codeTable}) {
     _socket.add(_generator.printCodeTable(codeTable: codeTable));
   }
 
@@ -121,12 +121,12 @@ class NetworkPrinter {
   }
 
   void imageRaster(
-    Image image, {
-    PosAlign align = PosAlign.center,
-    bool highDensityHorizontal = true,
-    bool highDensityVertical = true,
-    PosImageFn imageFn = PosImageFn.bitImageRaster,
-  }) {
+      Image image, {
+        PosAlign align = PosAlign.center,
+        bool highDensityHorizontal = true,
+        bool highDensityVertical = true,
+        PosImageFn imageFn = PosImageFn.bitImageRaster,
+      }) {
     _socket.add(_generator.imageRaster(
       image,
       align: align,
@@ -137,13 +137,13 @@ class NetworkPrinter {
   }
 
   void barcode(
-    Barcode barcode, {
-    int width,
-    int height,
-    BarcodeFont font,
-    BarcodeText textPos = BarcodeText.below,
-    PosAlign align = PosAlign.center,
-  }) {
+      Barcode barcode, {
+        int? width,
+        int? height,
+        BarcodeFont? font,
+        BarcodeText textPos = BarcodeText.below,
+        PosAlign align = PosAlign.center,
+      }) {
     _socket.add(_generator.barcode(
       barcode,
       width: width,
@@ -155,11 +155,11 @@ class NetworkPrinter {
   }
 
   void qrcode(
-    String text, {
-    PosAlign align = PosAlign.center,
-    QRSize size = QRSize.Size4,
-    QRCorrection cor = QRCorrection.L,
-  }) {
+      String text, {
+        PosAlign align = PosAlign.center,
+        QRSize size = QRSize.Size4,
+        QRCorrection cor = QRCorrection.L,
+      }) {
     _socket.add(_generator.qrcode(text, align: align, size: size, cor: cor));
   }
 
@@ -167,16 +167,16 @@ class NetworkPrinter {
     _socket.add(_generator.drawer(pin: pin));
   }
 
-  void hr({String ch = '-', int len, int linesAfter = 0}) {
+  void hr({String ch = '-', int? len, int linesAfter = 0}) {
     _socket.add(_generator.hr(ch: ch, linesAfter: linesAfter));
   }
 
   void textEncoded(
-    Uint8List textBytes, {
-    PosStyles styles = const PosStyles(),
-    int linesAfter = 0,
-    int maxCharsPerLine,
-  }) {
+      Uint8List textBytes, {
+        PosStyles styles = const PosStyles(),
+        int linesAfter = 0,
+        int? maxCharsPerLine,
+      }) {
     _socket.add(_generator.textEncoded(
       textBytes,
       styles: styles,
@@ -184,5 +184,5 @@ class NetworkPrinter {
       maxCharsPerLine: maxCharsPerLine,
     ));
   }
-  // ************************ (end) Printer Commands ************************
+// ************************ (end) Printer Commands ************************
 }
